requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
az-iranian-bank-gateways==1.6.18
beautifultable==0.8.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.4
Deprecated==1.2.13
Django==3.2.15
django-crispy-forms==1.14.0
django-decouple==2.1
django-jalali==6.0.0
django-redis==5.2.0
djangorestframework==3.13.1
fandogh-cli==1.26.10
idna==3.3
isodate==0.6.1
jdatetime==4.1.0
kavenegar==1.1.2
kombu==5.2.4
lxml==4.9.1
packaging==21.3
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pycparser==2.21
pycryptodome==3.15.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pypng==0.20220715.0
PyQRCode==1.2.1
pytz==2022.2.1
PyYAML==5.4.1
redis==4.3.4
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.9.1
retrying==1.3.3
six==1.16.0
sqlparse==0.4.2
tzlocal==2.1
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.3.3
wrapt==1.14.1
zeep==4.1.0
